requirements.txt
python-dotenv==1.2.1
chromadb==1.4.1
langchain-groq==1.1.1
langchain-community==0.4.1
langchain-text-splitters==1.1.0
selenium==4.40.0
streamlit==1.53.0
onnxruntime==1.23.2